+++ packages/react/src/experimental/AiChat/components/UserMessage.tsx
@@ -29,7 +29,7 @@ export const UserMessage = ({ message, ImageRenderer }: UserMessageProps) => {
   return (
     <div
       ref={ref}
-      className="my-4 w-fit max-w-[330px] self-end whitespace-pre-wrap rounded-2xl border border-solid border-f1-border-secondary bg-f1-background-tertiary px-4 py-3 first:mt-0 last:mb-0"
+      className="my-4 w-fit max-w-[min(90%,330px)] self-end whitespace-pre-wrap rounded-2xl border border-solid border-f1-border-secondary bg-f1-background-tertiary px-4 py-3 first:mt-0 last:mb-0"
     >
       {message?.content}
     </div>