+++ packages/react/src/experimental/AiChat/exports.ts
@@ -1,13 +1,8 @@
 export {
-  AiChat,
-  AiChatProvider,
-  type AiChatProps,
-  type AiChatProviderProps,
-} from "./index"
-
-export {
   HILActionConfirmation,
   type HILActionConfirmationProps,
 } from "./HILActionConfirmation"
+export { AiChat, AiChatProvider, type AiChatProviderProps } from "./index"
+export { useAiChat } from "./providers/AiChatStateProvider"
 
 export { ActionItem, type ActionItemProps } from "./ActionItem"